updated springbboot petstore implicit header sample

+++ samples/server/petstore/springboot-implicitHeaders/src/test/java/io/swagger/api/PetApiControllerIntegrationTest.java
@@ -32,7 +32,7 @@ public class PetApiControllerIntegrationTest {
 
     @Test
     public void deletePetTest() throws Exception {
-        Integer petId = 56;
+        Long petId = 56L;
         ResponseEntity<Void> responseEntity = api.deletePet(petId);
         assertEquals(HttpStatus.NOT_IMPLEMENTED, responseEntity.getStatusCode());
     }
@@ -53,7 +53,7 @@ public class PetApiControllerIntegrationTest {
 
     @Test
     public void getPetByIdTest() throws Exception {
-        Integer petId = 56;
+        Long petId = 56L;
         ResponseEntity<Pet> responseEntity = api.getPetById(petId);
         assertEquals(HttpStatus.NOT_IMPLEMENTED, responseEntity.getStatusCode());
     }
@@ -67,7 +67,7 @@ public class PetApiControllerIntegrationTest {
 
     @Test
     public void updatePetWithFormTest() throws Exception {
-        Integer petId = 56;
+        Long petId = 56L;
         String name = "name_example";
         String status = "status_example";
         ResponseEntity<Void> responseEntity = api.updatePetWithForm(petId, name, status);
@@ -76,7 +76,7 @@ public class PetApiControllerIntegrationTest {
 
     @Test
     public void uploadFileTest() throws Exception {
-        Integer petId = 56;
+        Long petId = 56L;
         String additionalMetadata = "additionalMetadata_example";
         org.springframework.web.multipart.MultipartFile file = null;
         ResponseEntity<ModelApiResponse> responseEntity = api.uploadFile(petId, additionalMetadata, file);
